requirements.txt
aiosqlite==0.20.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
blessings==1.7
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colour-runner==0.1.1
coverage==7.4.4
cryptography==42.0.5
deepdiff==7.0.1
distlib==0.3.8
ecdsa==0.19.0
fastapi==0.110.1
filelock==3.13.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
iniconfig==2.0.0
install==1.3.5
loguru==0.7.2
Mako==1.3.3
MarkupSafe==2.1.5
mypy==1.9.0
mypy-extensions==1.0.0
ordered-set==4.1.0
packaging==24.0
passlib==1.7.4
platformdirs==4.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
Pygments==2.17.2
PyJWT==1.7.1
pyproject-api==1.6.1
pytest==8.1.1
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.31.0
rootpath==0.1.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
sqlalchemyseed==2.0.0
starlette==0.37.2
termcolor==2.4.0
tox==4.14.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.25.1
